fix(ai-gateway,studio): remove import cycle in fetch credits
